Move old config to root level
