+++ packages/eslint-config-expo/package.json
@@ -19,6 +19,7 @@
   "files": [
     "README.md",
     "default.js",
+    "flat",
     "utils"
   ],
   "author": "Expo",
