+++ packages/eslint-config-expo/utils/core.js
