+++ packages/eslint-config-expo/utils/extensions.js
