+++ packages/eslint-config-expo/utils/typescript.js
